+++ scandir_linux.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/immortal/xtime"
@@ -17,7 +18,7 @@ import (
 type ScanDir struct {
 	scandir       string
 	sdir          string
-	services      map[string]string
+	services      sync.Map
 	timeMultipler time.Duration
 }
 
@@ -51,7 +52,6 @@ func NewScanDir(path string) (*ScanDir, error) {
 	return &ScanDir{
 		scandir:       dir,
 		sdir:          GetSdir(),
-		services:      map[string]string{},
 		timeMultipler: 5,
 	}, nil
 }
@@ -92,14 +92,12 @@ func (s *ScanDir) Scanner(ctl Control) {
 			}
 			// add service to services map or reload if file has been changed
 			services = append(services, name)
-			if hash, ok := s.services[name]; !ok || !isFile(filepath.Join(s.sdir, name, "lock")) {
-				s.services[name] = md5
+			if hash, ok := s.services.Load(name); !ok || !isFile(filepath.Join(s.sdir, name, "lock")) {
 				start = true
 			} else if hash != md5 {
-				// update to new hash
-				s.services[name] = md5
 				stop = true
 			}
+			s.services.Store(name, md5)
 			// check if file hasn't been changed since last tick (5 seconds)
 			refresh := (time.Now().Unix() - xtime.Get(f).Ctime().Unix()) <= int64(s.timeMultipler)
 			if refresh || start {
@@ -113,7 +111,7 @@ func (s *ScanDir) Scanner(ctl Control) {
 				if _, err := ctl.SendSignal(filepath.Join(s.sdir, name, "immortal.sock"), "start"); err != nil {
 					if out, err := ctl.Run(fmt.Sprintf("immortal -c %s -ctl %s", path, name)); err != nil {
 						// keep retrying
-						delete(s.services, name)
+						s.services.Delete(name)
 						log.Println(err)
 					} else {
 						log.Printf("%s\n", out)
@@ -133,7 +131,7 @@ func (s *ScanDir) Scanner(ctl Control) {
 	// halts services that don't exist anymore
 	for service := range s.services {
 		if !inSlice(services, service) {
-			delete(s.services, service)
+			s.services.Delete(service)
 			ctl.SendSignal(filepath.Join(s.sdir, service, "immortal.sock"), "halt")
 			log.Printf("Exiting: %s\n", service)
 		}