+++ CHANGELOG.md
@@ -5,5 +5,5 @@
 * Improved lint
 * Print cmd name (not just PID) in the log when the process terminates [#29](https://github.com/immortal/immortal/pull/29)
 * Removed info.go (signal.Notify) from supervise.go
-* Replaced lock/map with sync.Map in scandir.go
+* Replaced lock/map with sync.Map in scandir*.go
 * Updated HandleSignal to use `GetParam` from violetear